Fix resizing bug when toggling visibility
The Pad is now responsible for know which view to adjust to
